FINALLY Figure out webpack
Nominally, this commit is where I add loaders
  - sass-loader to handle parsing sass files to css
  - css-loader to handle parsing css files to JS
  - style-loader to inject said styles into HTML
  - HtmlWebpackPlugin to handle injecting the renderer bundle into Electron

But really, this commit was for me to figure out what Webpack actually does and
how it does it. To that end, I've now structured the app closer to how I think I
want it to be:

"Main" concerns live at the root, "Renderer" concerns live in "/renderer", and
mostly importantly, the App will live in "/src".  Webpack will bundle into
"renderer/", and also owns converting the template index.pug into the rendered
index.html

The dist/ folder is deleted, as that was a holdover from web-thinking.