Inflate Goal as Possible in Reducer
Previously, the app required a goal to have all elements filled in to be an
acceptable goal for the store. This wasn't a mathematical requirement though- a
goal can have its final attribute calculated if given the others.

This adds the ability to calculate goalTotal, deadlineYear, or spendingPerMonth
in the goals reducer. The logic is very close to that required on updating a
goal, and may suggest a future refactor.